requirements/base.txt
Django>=5.2
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary
python-dotenv
celery
redis
django-redis